website server

+++ JumpScale9Lib/tools/docgenerator/DocSite.py
@@ -1,8 +1,6 @@
 from js9 import j
 from .Doc import Doc
-from .DocBase import DocBase
 from .DocWatchdog import DocWatchdog
-from .HtmlPage import HtmlPage
 JSBASE = j.application.jsbase_get_class()
 
 import copy
@@ -65,7 +63,7 @@ class DocSite(JSBASE):
         self.others = {}
         self.files = {}
         self.sidebars = {}
-        self._processed = False
+        
 
     
         # check if there are dependencies
@@ -82,7 +80,7 @@ class DocSite(JSBASE):
         self.logger.level=1
 
         self._git=None
-        self._watchdog=None
+        self._loaded = False
 
         self.logger.info("loaded:%s"%self)
         
@@ -138,7 +136,6 @@ class DocSite(JSBASE):
         rdirpath = rdirpath.strip("/").strip().strip("/")
         self.data_default[rdirpath] = data
 
-
     def load(self):
         """
         walk in right order over all files which we want to potentially use (include)
@@ -147,6 +144,8 @@ class DocSite(JSBASE):
         if duplicate only the first found will be used
 
         """
+        if self._loaded:
+            return
 
         j.sal.fs.remove(self.path + "errors.md")
         path = self.path
@@ -200,12 +199,13 @@ class DocSite(JSBASE):
                 self.docs[doc.name_dot_lower] = doc
             elif ext in ["html","htm"]:
                 self.logger.debug("found html:%s"%path)
-                l = len(ext)+1
-                base = base[:-l]  # remove extension
-                doc = HtmlPage(path, base, docsite=self)
-                # if base not in self.htmlpages:
-                #     self.htmlpages[base.lower()] = doc
-                self.htmlpages[doc.name_dot_lower] = doc
+                raise RuntimeError()
+                # l = len(ext)+1
+                # base = base[:-l]  # remove extension
+                # doc = HtmlPage(path, base, docsite=self)
+                # # if base not in self.htmlpages:
+                # #     self.htmlpages[base.lower()] = doc
+                # self.htmlpages[doc.name_dot_lower] = doc
             else:
                 
                 if ext in ["png", "jpg", "jpeg", "pdf", "docx", "doc", "xlsx", "xls", \
@@ -235,6 +235,8 @@ class DocSite(JSBASE):
             callbackForMatchDir=callbackForMatchDir,
             callbackForMatchFile=callbackForMatchFile)
 
+        self._loaded=True
+
     # def file_add(self, path):
     #     if not j.sal.fs.exists(path, followlinks=True):
     #         raise j.exceptions.Input(message="Cannot find path:%s" % path, level=1, source="", tags="", msgpub="")
@@ -252,10 +254,10 @@ class DocSite(JSBASE):
     #     for name, path in self.files.items():
     #         j.sal.fs.copyFile(path, j.sal.fs.joinPaths(dpath, name))
 
-    def process(self):
-        for key, doc in self.docs.items():
-            doc.process()
-        self._processed = True
+    # def process(self):
+    #     for key, doc in self.docs.items():
+    #         doc.process()
+    #     self._processed = True
 
     def error_raise(self, errormsg, doc=None):
         if doc is not None:
@@ -269,13 +271,9 @@ class DocSite(JSBASE):
             embed()
             raise RuntimeError("stop debug here")
 
-    def load_process(self):
-        self.load()
-        self.process()
 
     def file_get(self, name, die=True):
-        if not self._processed:
-            self.load_process()
+        self.load()
         
         for key, val in self.files.items():
             if key.lower() == name.lower():
@@ -287,8 +285,9 @@ class DocSite(JSBASE):
 
     def doc_get(self, name, cat="", die=True):
         
-        if not self._processed:
-            self.load_process()
+        import pudb; pudb.set_trace()
+        
+        self.load()
 
         if j.data.types.list.check(name):
             name = "/".join(name)
@@ -360,29 +359,13 @@ class DocSite(JSBASE):
             if len(res)>0:
                 return  len(res),self.docs[res[0]]
             else:
-                return 0,None
-
-
-    def html_get(self, name, die=True):
-        if not self._processed:
-            self.load_process()        
-        if "." in name:
-            ext =  j.sal.fs.getFileExtension(name).lower()
-            name = name[:-(len(ext)+1)] #name without extension
-        name = j.data.text.strip_to_ascii_dense(name)
-        if name in self.htmlpages:
-            return self.htmlpages[name]
-        if die:
-            raise j.exceptions.Input(message="Cannot find doc with name:%s" % name, level=1, source="", tags="", msgpub="")
-        else:
-            return None       
+                return 0,None  
 
     def sidebar_get(self, url):
         """
         will calculate the sidebar, if not in url will return None
         """
-        if not self._processed:
-            self.load_process()        
+        self.load()        
         if j.data.types.list.check(url):
             url = "/".join(url)
         self.logger.debug("sidebar_get:%s"%url)
